Use the socket factory for direct connections as well.
Currently, the passed-in socket factory is only used for
connections to HTTP proxies. I think this was not the intent of
the original socket factory change, because the commit message
said that the "socket factory will be used for all non-proxy
connections and HTTP proxy connections".

So use it for DIRECT connections as well.

Change-Id: Ia77dea817be11a0199a3dda4b4c7c90d59a9030c