Avoid a reverse DNS-lookup for a numeric proxy address
When the proxy is specified as an IP address the RouteSelector
would force a reverse-lookup. This can introduce extra latency
when the DNS is slow or missing. It also then turns the name
back into a set of IPs, which may result in the original IP
address not being given priority.

Bug: 18327075
Change-Id: Idafd72f4d5b901e267081f924377db4ada9afe1b