[FOR DISCUSSION] Fix to com.audible.
The authentication scheme that this app uses is fairly involved. They
compute a SHA digest of a whole bunch of things including the request
URL and send that along with their request. This is presumably verified
at the server. The issue is :

(1) the URL (or URI) that the app thinks it's sending a request to
contains the device model.
(2) the device model can contain a space ("Nexus 6" for example)
(3) the encoded java.net.URL (or URI) query looks something like
"/foo?model=Nexus+6".
(4) on the okhttp side it becomes "/foo?model=Nexus%206" because HttpUrl
prefers to use "%20" instead of + to avoid special casing. (a rational
choice, IMO..)

This is compliant with the spec but not with their authentication scheme.

Change-Id: Ic47f45d7d7f88aae935fcf5e6230a9e9924a67c7